REF: Base class for array
