+++ cyberpandas/__init__.py
@@ -4,7 +4,6 @@ from .ip_array import (  # noqa
     IPType,
     IPArray,
     IPAccessor,
-    IPAddressIndex,
 )
 from .parser import to_ipaddress  # noqa
